requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic[email]==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
# passlib 1.7.4 is incompatible with bcrypt>=4.0 (removed __about__ + strict 72-byte limit).
# Pin to the last compatible 3.x release.
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
